journal-app: replace auto-resize textarea library w/ local solution; remove unused dependencies
I tried to remove the date selection and scrolling dependencies as well,
but local solutions ended up increasing the code complexity to a degree
that made the effort feel unprofitable.

In the case of date selection, modern HTML provides a robust and
user-friendly solution in the form of `<input type="date"/>`.
Unfortunately, this mechanism is a bit tricky to use as a controlled
component in React and requires a nontrivial amount of adjustment
to account for timezones.

In the case of scrolling, an alternate home-grown solution is feasible
using React's `useRef` and Javascript's `IntersectionObserver`; however,
the simple solution is a bit buggy and anything approximating parity with
`react-bottom-scroll-listener` requires 4-5 extra state variables and
additional complexity surrounding handling of updates and loading.